consolidated space groups into xrsdkit.definitions
